[Packages] Rename ZBPackageDepictionViewController -> ZBPackageViewController
