+++ Zebra/Tabs/Home/ZBHomeTableViewController.m
@@ -8,7 +8,7 @@
 
 #import "ZBHomeTableViewController.h"
 #import "ZBNewsCollectionViewCell.h"
-#import "ZBPackageDepictionViewController.h"
+#import "ZBPackageViewController.h"
 
 #import <ZBSettings.h>
 #import <Tabs/Home/Credits/ZBCreditsTableViewController.h>
@@ -50,7 +50,7 @@ typedef enum ZBLinksOrder : NSUInteger {
     NSMutableArray *redditPosts;
     BOOL hideUDID;
 }
-@property (nonatomic, weak) ZBPackageDepictionViewController *previewPackageDepictionVC;
+@property (nonatomic, weak) ZBPackageViewController *previewPackageDepictionVC;
 @end
 
 @implementation ZBHomeTableViewController
@@ -665,7 +665,7 @@ typedef enum ZBLinksOrder : NSUInteger {
     ZBPackage *package = [[ZBDatabaseManager sharedInstance] topVersionForPackageID:cell.packageID];
     
     if (package) {
-        ZBPackageDepictionViewController *packageDepiction = [[ZBPackageDepictionViewController alloc] initWithPackage:package];
+        ZBPackageViewController *packageDepiction = [[ZBPackageViewController alloc] initWithPackage:package];
         [[self navigationController] pushViewController:packageDepiction animated:YES];
     }
 }
